feat: put switches under Edgenote preview to disable parts of expansion
