+++ app/assets/stylesheets/blueprint_overrides.sass
@@ -20,3 +20,7 @@
 .pt-toast-message
   .pt-progress-bar
     margin-top: 5px
+
+.pt-control
+  input
+    position: absolute !important
