+++ config/locales/en.yml
@@ -348,6 +348,9 @@ en:
       preview: Edgenote preview
       pull_a_quote: Pull a quote
       remove: Remove
+      use_embed: Embed a widget from this website
+      use_preview_description: Include a description in the link preview
+      use_preview_image: Include an image in the the link preview
       upload_an_audio_file: Upload an audio file
       write_a_caption: Write a caption
     index:
